FIX: Speed bonus would always show as 0 whenever the speed bonus multiplier factor was < 100 FIX: Crashes on loading and unloading when the speed bonus multiplier factor was < 100 FIX: Crashes relating to the GUI for showing goods/passengers on vehicles or at stations FIX: Minor optimisation of the routing code (Knightly) CHANGE: "No route" colour is now a lighter blue to make it easier to read the text, "no route"

+++ dataobj/einstellungen.cc
@@ -265,7 +265,7 @@ einstellungen_t::einstellungen_t() :
 
 	// Global power factor
 	// @author: jamespetts
-	global_power_factor = 1.0;
+	global_power_factor = 1.0F;
 
 	avoid_overcrowding = false;
 	
@@ -277,7 +277,7 @@ einstellungen_t::einstellungen_t() :
 	// @author: jamespetts
 	enforce_weight_limits = 1;
 
-	speed_bonus_multiplier = 1.0;
+	speed_bonus_multiplier = 1.0F;
 }
 
 
@@ -722,7 +722,7 @@ void einstellungen_t::rdwr(loadsave_t *file)
 			file->rdwr_byte(enforce_weight_limits, "");
 			uint16 speed_bonus_multiplier_percent = speed_bonus_multiplier * 100;
 			file->rdwr_short(speed_bonus_multiplier_percent, "");
-			speed_bonus_multiplier = (float)speed_bonus_multiplier_percent / 100;
+			speed_bonus_multiplier = (float)speed_bonus_multiplier_percent / 100.0F;
 		}
 
 	}
@@ -1017,7 +1017,7 @@ void einstellungen_t::parse_simuconf( tabfile_t &simuconf, sint16 &disp_width, s
 
 	uint16 speed_bonus_multiplier_percent = 100;
 	speed_bonus_multiplier_percent = contents.get_int("speed_bonus_multiplier_percent", speed_bonus_multiplier_percent);
-	speed_bonus_multiplier = speed_bonus_multiplier_percent / 100;
+	speed_bonus_multiplier = (float)speed_bonus_multiplier_percent / 100.0F;
 
 	/*
 	 * Selection of savegame format through inifile
